03/06/24

+++ src/basic_solidity/SimpleStorage.sol
@@ -3,12 +3,11 @@
 
  contract SimpleStorage {
     uint256 public num;
-    
-    function set(uint256 _num) public {
+    function set(uint _num) public {
         num = _num;
     }
 
-    function get() public view returns (uint256) {
+    function get() public view returns (uint) {
         return num;
     }
 }
